rename redis
