+++ cache/redis/redis_test.go
@@ -1,4 +1,4 @@
-package rediscache
+package redis
 
 import (
 	"context"
